+++ lib/src/platform/hawcx_flutter_sdk_platform.dart
@@ -58,4 +58,3 @@ abstract class HawcxFlutterSdkPlatform extends PlatformInterface {
 
   Future<void> declinePushRequest(String requestId);
 }
-